Added support for iOS specific features
